moved add aliases to util function
