+++ src/utils/add-aliases.js
@@ -1,0 +1,11 @@
+const addAliases = (arr, aliases) => 
+  aliases.forEach((key, i) =>
+    Object.defineProperty(arr, key, {
+      enumerable: false,
+      get() {
+        return this[i]
+      },
+    })
+  )
+
+export default addAliases
